+++ src/modules/mavlink/mavlink_messages.cpp
@@ -134,6 +134,7 @@ using matrix::wrap_2pi;
 #include "streams/SERVO_OUTPUT_RAW.hpp"
 #include "streams/STATUSTEXT.hpp"
 #include "streams/STORAGE_INFORMATION.hpp"
+#include "streams/SYSTEM_TIME.hpp"
 #include "streams/TRAJECTORY_REPRESENTATION_WAYPOINTS.hpp"
 #include "streams/VIBRATION.hpp"
 #include "streams/WIND_COV.hpp"
@@ -1592,68 +1593,6 @@ protected:
 };
 #endif
 
-class MavlinkStreamSystemTime : public MavlinkStream
-{
-public:
-	const char *get_name() const override
-	{
-		return MavlinkStreamSystemTime::get_name_static();
-	}
-
-	static constexpr const char *get_name_static()
-	{
-		return "SYSTEM_TIME";
-	}
-
-	static constexpr uint16_t get_id_static()
-	{
-		return MAVLINK_MSG_ID_SYSTEM_TIME;
-	}
-
-	uint16_t get_id() override
-	{
-		return get_id_static();
-	}
-
-	static MavlinkStream *new_instance(Mavlink *mavlink)
-	{
-		return new MavlinkStreamSystemTime(mavlink);
-	}
-
-	unsigned get_size() override
-	{
-		return MAVLINK_MSG_ID_SYSTEM_TIME_LEN + MAVLINK_NUM_NON_PAYLOAD_BYTES;
-	}
-
-private:
-	/* do not allow top copying this class */
-	MavlinkStreamSystemTime(MavlinkStreamSystemTime &) = delete;
-	MavlinkStreamSystemTime &operator = (const MavlinkStreamSystemTime &) = delete;
-
-protected:
-	explicit MavlinkStreamSystemTime(Mavlink *mavlink) : MavlinkStream(mavlink)
-	{}
-
-	bool send() override
-	{
-		timespec tv;
-		px4_clock_gettime(CLOCK_REALTIME, &tv);
-
-		mavlink_system_time_t msg{};
-		msg.time_boot_ms = hrt_absolute_time() / 1000;
-		msg.time_unix_usec = (uint64_t)tv.tv_sec * 1000000 + tv.tv_nsec / 1000;
-
-		// If the time is before 2001-01-01, it's probably the default 2000
-		// and we don't need to bother sending it because it's definitely wrong.
-		if (msg.time_unix_usec > 978307200000000) {
-			mavlink_msg_system_time_send_struct(_mavlink->get_channel(), &msg);
-			return true;
-		}
-
-		return false;
-	}
-};
-
 class MavlinkStreamTimesync : public MavlinkStream
 {
 public:
@@ -1923,7 +1862,9 @@ static const StreamListItem streams_list[] = {
 #endif // GPS_GLOBAL_ORIGIN_HPP
 	create_stream_list_item<MavlinkStreamGPSRawInt>(),
 	create_stream_list_item<MavlinkStreamGPS2Raw>(),
+#if defined(SYSTEM_TIME_HPP)
 	create_stream_list_item<MavlinkStreamSystemTime>(),
+#endif // SYSTEM_TIME_HPP
 	create_stream_list_item<MavlinkStreamTimesync>(),
 #if defined(GLOBAL_POSITION_INT_HPP)
 	create_stream_list_item<MavlinkStreamGlobalPositionInt>(),
